Update zh-cn.lng
The original translation was done by a machine, and the quality was very poor. Many features were hard to understand. The version submitted this time is a polished Simplified Chinese translation. Additionally, the language file does not work in some interfaces (for example, the entire control bar of KVM), and some controls still display in English. The language file does include this part of the content, which needs to be resolved by Stardock.